% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books.R
\docType{data}
\name{ashorthistoryofengland}
\alias{ashorthistoryofengland}
\title{The text of G.K. Chesterton's "A Short History of England"}
\format{A character vector with 5367 elements}
\source{
\url{http://www.gutenberg.org/ebooks/20897}
}
\usage{
ashorthistoryofengland
}
\description{
A dataset containing the text of G.K. Chesterton's 
"A Short History of England", 
published in 1917.
The UTF-8 plain text was sourced from Project Gutenberg
and is divided into elements of up to about 70 characters each.
(Some elements are blank.)
}
\keyword{datasets}
