% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chesterton_books.R
\name{chesterton_books}
\alias{chesterton_books}
\title{Tidy data frame of G.K. Chesterton's 6 works.}
\usage{
chesterton_books()
}
\value{
A data frame with two columns: \code{text} and \code{book}
}
\description{
Returns a tidy data frame of G.K. Chesterton's 6 completed, published novels with 
two columns: \code{text}, which contains the text of the novels divided into 
elements of up to about 70 characters each, and \code{book}, which contains the titles of
the novels as a factor in order of publication.
}
\details{
Users should be aware that there are some differences in usage 
between the novels as made available by Project Gutenberg. For example, 
"anything" vs. "any thing", "Mr" vs. "Mr.", and using underscores vs. all 
caps to indicate italics/emphasis.
}
\examples{

library(dplyr)

chesterton_books() \%>\% group_by(book) \%>\%
     summarise(total_lines = n())

}
