% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books.R
\docType{package}
\name{gkchestertonr}
\alias{gkchestertonr}
\alias{gkchestertonr-package}
\title{G.K. Chesterton's Complete Works.}
\description{
This package contains the complete text of G.K. Chesterton's works, 
formatted to be convenient for text analysis.
}
