% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books.R
\docType{data}
\name{manalive}
\alias{manalive}
\title{The text of G.K. Chesterton's novel "Man Alive"}
\format{A character vector with 6336 elements}
\source{
\url{http://www.gutenberg.org/ebooks/1718}
}
\usage{
manalive
}
\description{
A dataset containing the text of G.K. Chesterton's 1912 novel "Man Alive". 
The UTF-8 plain text was sourced from Project Gutenberg
and is divided into elements of up to about 70 characters each.
(Some elements are blank.)
}
\keyword{datasets}
