% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books.R
\docType{data}
\name{theballadofthewhitehorse}
\alias{theballadofthewhitehorse}
\title{The text of G.K. Chesterton's "The Ballad of the White Horse"}
\format{A character vector with 3341 elements}
\source{
\url{http://www.gutenberg.org/ebooks/1719}
}
\usage{
theballadofthewhitehorse
}
\description{
A dataset containing the text of G.K. Chesterton's  
"The Ballad of the White Horse", published in 1911.
The UTF-8 plain text was sourced from Project Gutenberg
and is divided into elements of up to about 70 characters each.
(Some elements are blank.)
}
\keyword{datasets}
