% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books.R
\docType{data}
\name{themanwhoknewtoomuch}
\alias{themanwhoknewtoomuch}
\title{The text of G.K. Chesterton's "The Man Who Knew too much"}
\format{A character vector with 6449 elements}
\source{
\url{http://www.gutenberg.org/ebooks/1720}
}
\usage{
themanwhoknewtoomuch
}
\description{
A dataset containing the text of G.K. Chesterton's detective story 
"The Man Who Knew too much", 
published in 1922.
The UTF-8 plain text was sourced from Project Gutenberg
and is divided into elements of up to about 70 characters each.
(Some elements are blank.)
}
\keyword{datasets}
