% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books.R
\docType{data}
\name{themanwhowasthursday}
\alias{themanwhowasthursday}
\title{The text of G.K. Chesterton's novel The Man Who Was Thursday: A Nightmare}
\format{A character vector with 6462 elements}
\source{
\url{http://www.gutenberg.org/ebooks/1695}
}
\usage{
themanwhowasthursday
}
\description{
A dataset containing the text of G.K. Chesterton's novel 
"The Man Who Was Thursday: A Nightmare", 
first published in 1908
The UTF-8 plain text was sourced from Project Gutenberg
and is divided into elements of up to about 70 characters each.
(Some elements are blank.)
}
\keyword{datasets}
