% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/novels.R
\docType{data}
\name{thenapoleonofnothinghill}
\alias{thenapoleonofnothinghill}
\title{The text of G.K. Chesterton's novel "The Napoleon of Nothing Hill"}
\format{A character vector with xxx elements}
\source{
\url{http://www.gutenberg.org/ebooks/20058}
}
\usage{
thenapoleonofnothinghill
}
\description{
A dataset containing the text of G.K. Chesterton's novel 
"The Napoleon of Nothing Hill", published in 1904.
The UTF-8 plain text was sourced from Project Gutenberg
and is divided into elements of up to about 70 characters each.
(Some elements are blank.)
}
\keyword{datasets}
