% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/novels.R
\docType{data}
\name{themanwhowasthursday}
\alias{themanwhowasthursday}
\title{The text of G.K. Chesterton's novel "The Man Who Was Thursday"}
\format{A character vector with xxx elements}
\source{
\url{http://www.gutenberg.org/ebooks/xxx}
}
\usage{
themanwhowasthursday
}
\description{
A dataset containing the text of G.K. Chesterton's novel "The Man Who Was Thursday", 
published posthumously in 1818.
The UTF-8 plain text was sourced from Project Gutenberg
and is divided into elements of up to about 70 characters each.
(Some elements are blank.)
}
\keyword{datasets}
