% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/novels.R
\docType{data}
\name{thenewjerusalem}
\alias{thenewjerusalem}
\title{The text of G.K. Chesterton's novel ""}
\format{A character vector with xxx elements}
\source{
\url{http://www.gutenberg.org/ebooks/158}
}
\usage{
thenewjerusalem
}
\description{
A dataset containing the text of G.K. Chesterton's 1815 novel "Emma".
The UTF-8 plain text was sourced from Project Gutenberg
and is divided into elements of up to about 70 characters each. 
(Some elements are blank.)
}
\keyword{datasets}
